Renamed to BukkitTelnet
