+++ README
@@ -1,0 +1,5 @@
+BukkitTelnet is a Telnet server for Minecraft. It allows you to remotely administrate your server via any Telnet client.
+
+BukkitTelnet is a fork of MCTelnet that is designed for servers with multiple administrators who all have the same full permissions. It allows for login under any username provided that the proper password is supplied.
+
+MCTelnet was originally created by bekvon. I modified it heavily to be better suited to my TotalFreedomServer, and eventually renamed it.
