+++ nbproject/private/private.properties
@@ -3,4 +3,4 @@ do.depend=false
 do.jar=true
 javac.debug=true
 javadoc.preview=true
-user.properties.file=C:\\Users\\Steven\\.netbeans\\7.0\\build.properties
+user.properties.file=C:\\Users\\Steven\\.netbeans\\7.1.1\\build.properties
