+++ nbproject/project.properties
@@ -2,7 +2,7 @@ annotation.processing.enabled=true
 annotation.processing.enabled.in.editor=false
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-application.title=MCTelnet
+application.title=BukkitTelnet
 application.vendor=Administrator
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
@@ -22,7 +22,7 @@ debug.test.classpath=\
     ${run.test.classpath}
 # This directory is removed when the project is cleaned:
 dist.dir=dist
-dist.jar=${dist.dir}/MCTelnet.jar
+dist.jar=${dist.dir}/BukkitTelnet.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 endorsed.classpath=
 excludes=
